# docs/swagger.yaml
basePath: /
definitions:
...
  lib.ListResponse:
    properties:
      message:
        type: string
      page_info:
        $ref: '#/definitions/lib.PageInfo'
      result: {}
      status:
        type: integer
    type: object
  lib.Movie:
    properties:
      caster_id:
        description: Legacy
        type: integer
      casters:
        items:
          type: string
        type: array
      director_name:
        type: string
      duration:
        type: string
      genre_id:
        description: 'Legacy: keeping for now but will use Genres'
        type: integer
      genres:
        items:
          type: string
        type: array
      id:
        type: integer
      image:
        type: string
      recommendation:
        type: boolean
      released_at:
        type: string
      synopsis:
        type: string
      title:
        type: string
    type: object
  lib.MovieShowtime:
    properties:
      cinema_id:
        type: integer
      id:
        type: integer
      movie_id:
        type: integer
      price:
        type: integer
      show_date:
        type: string
      show_time:
        type: string
    type: object
...
  lib.PageInfo:
    properties:
      current_page:
        type: integer
      next_page:
        type: integer
      prev_page:
        type: integer
      total_data:
        type: integer
      total_page:
        type: integer
    type: object
...
  lib.Response:
    properties:
      message:
        type: string
      result: {}
      status:
        type: integer
    type: object
...
host: localhost:8888
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for Movie Ticket Booking System
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Tikitz API
  version: "1.0"
paths:
...
  /movies:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of movies with optional filters
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      - description: Search by title
        in: query
        name: search
        type: string
      - description: Sort order (asc/desc)
        in: query
        name: sort
        type: string
      - description: Filter by release month
        in: query
        name: month
        type: integer
      - description: Filter by release year
        in: query
        name: year
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/lib.ListResponse'
            - properties:
                result:
                  items:
                    $ref: '#/definitions/lib.Movie'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Get all movies
      tags:
      - movies
    post:
      consumes:
      - multipart/form-data
      description: Insert a new movie into the database with its many-to-many relations
        and showtimes
      parameters:
      - description: Movie Title
        in: formData
        name: title
        required: true
        type: string
      - description: Release Date (RFC3339)
        in: formData
        name: released_at
        type: string
      - description: Is Recommended
        in: formData
        name: recommendation
        type: boolean
      - description: Duration
        in: formData
        name: duration
        type: string
      - description: Synopsis
        in: formData
        name: synopsis
        type: string
      - description: Director Name
        in: formData
        name: director_name
        type: string
      - collectionFormat: csv
        description: Genre IDs
        in: formData
        items:
          type: integer
        name: genre_ids
        type: array
      - collectionFormat: csv
        description: Caster IDs
        in: formData
        items:
          type: integer
        name: caster_ids
        type: array
      - collectionFormat: csv
        description: Cinema IDs
        in: formData
        items:
          type: integer
        name: cinema_ids
        type: array
      - description: Movie Poster
        in: formData
        name: image
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/lib.Response'
            - properties:
                result:
                  $ref: '#/definitions/lib.Movie'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - BearerAuth: []
      summary: Create a new movie
      tags:
      - movies
  /movies/{id}:
    delete:
      consumes:
      - application/json
      description: Remove a movie from the database by ID
      parameters:
      - description: Movie ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - BearerAuth: []
      summary: Delete movie
      tags:
      - movies
    get:
      consumes:
      - application/json
      description: Retrieve detailed information about a single movie
      parameters:
      - description: Movie ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/lib.Response'
            - properties:
                result:
                  $ref: '#/definitions/lib.Movie'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Get movie by ID
      tags:
      - movies
    patch:
      consumes:
      - multipart/form-data
      description: Update movie details by ID
      parameters:
      - description: Movie ID
        in: path
        name: id
        required: true
        type: integer
      - description: Movie Title
        in: formData
        name: title
        type: string
      - description: Release Date (RFC3339)
        in: formData
        name: released_at
        type: string
      - description: Is Recommended
        in: formData
        name: recommendation
        type: boolean
      - description: Duration
        in: formData
        name: duration
        type: string
      - description: Synopsis
        in: formData
        name: synopsis
        type: string
      - description: Director Name
        in: formData
        name: director_name
        type: string
      - collectionFormat: csv
        description: Genre IDs
        in: formData
        items:
          type: integer
        name: genre_ids
        type: array
      - collectionFormat: csv
        description: Caster IDs
        in: formData
        items:
          type: integer
        name: caster_ids
        type: array
      - collectionFormat: csv
        description: Cinema IDs
        in: formData
        items:
          type: integer
        name: cinema_ids
        type: array
      - description: Movie Poster
        in: formData
        name: image
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/lib.Response'
            - properties:
                result:
                  $ref: '#/definitions/lib.Movie'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - BearerAuth: []
      summary: Update movie
      tags:
      - movies
  /movies/{id}/showtimes:
    get:
      consumes:
      - application/json
      description: Retrieve showtimes for a specific movie, optionally filtered by
        location
      parameters:
      - description: Movie ID
        in: path
        name: id
        required: true
        type: integer
      - description: Filter by Location ID
        in: query
        name: location_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/lib.Response'
            - properties:
                result:
                  items:
                    $ref: '#/definitions/lib.MovieShowtime'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Get movie showtimes
      tags:
      - movies
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"